feat: extract layers into a single folder
