+++ src/fs-gen/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 clap = { version = "4.5.3", features = ["derive", "wrap_help", "string"] }
+dircpy = "0.3.16"
 fuse-backend-rs = "0.12.0"
 once_cell = "1.19.0"
 regex = "1.10.4"
@@ -14,4 +15,3 @@ serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.115"
 signal-hook = "0.3.17"
 validator = { version = "0.17.0", features = ["derive"] }
-vfs = "0.12.0"
